add dataloader creation function
