+++ cyto_dl/utils/__init__.py
@@ -1,9 +1,9 @@
+from .array import create_dataloader, extract_array_predictions
 from .config import kv_to_dict, remove_aux_key
 from .pylogger import get_pylogger
 from .rich_utils import enforce_tags, print_config_tree
 from .template_utils import (
     close_loggers,
-    create_dataloader,
     extras,
     get_metric_value,
     instantiate_callbacks,
